requirements.txt
requests
flask
flask_pymongo
python-nmap
pymongo==3.4.0
